requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
alembic==1.13.2
bcrypt==4.0.1
